fix: migrate blog post item adjustments
swizzle smaller components after BlogPostItem has been refactored in beta 22
