+++ src/theme/BlogPostItem/Content/index.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="react" />
+import type { Props } from '@theme/BlogPostItem/Content';
+export default function BlogPostItemContent({ children, className, }: Props): JSX.Element;
