+++ src/theme/BlogPostItem/Content/styles.module.css
@@ -1,0 +1,6 @@
+.abstract {
+  font-weight: bold;
+  margin-top: 2rem;
+  margin-bottom: 2rem;
+  font-size: 1.1rem;
+}
