+++ src/theme/BlogPostItem/Header/index.d.ts
@@ -1,0 +1,2 @@
+/// <reference types="react" />
+export default function BlogPostItemHeader(): JSX.Element;
