--- src/theme/BlogPostItem/index.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="react" />
-import type { Props } from '@theme/BlogPostItem';
-export default function BlogPostItem(props: Props): JSX.Element;
