--- src/theme/BlogPostItem/styles.module.css
@@ -1,26 +1,0 @@
-.blogPostTitle {
-  font-size: 3rem;
-}
-
-/**
-  Blog post title should be smaller on smaller devices
-**/
-@media (max-width: 576px) {
-.blogPostTitle {
-  font-size: 2rem;
-  }
-}
-
-.blogPostData {
-  font-size: 0.9rem;
-}
-
-.blogPostDetailsFull {
-  flex-direction: column;
-}
-.blogPostAbstract {
-  font-weight: bold;
-  margin-top: 2rem;
-  margin-bottom: 2rem;
-  font-size: 1.1rem;
-}
